pi-extension: redistribute the PR-numbered test file into per-module suites
pr106-review.test.ts was named after the review event (meaningless post-merge)
and had accreted as a catch-all, duplicating a couple of guards already covered
in tool.test.ts. Extract its recording-tracer harness into a shared, non-shipped
test-support.ts and fold each test into the suite for the module it exercises:

- tool.test.ts    <- tool error-status / content-capture cases
- llm.test.ts     <- message_end usage + error status, request privacy gate,
                     after_provider_response (status, rate-limit headers, errors)
- turn.test.ts    <- agent_start input/raw-input gate, async repo-slug attach
- session.test.ts <- provider lifecycle, per-session reset, raceWithTimeout
- command.test.ts <- /traceroot status|flush|disable|enable

Drop the two duplicated tool guards, delete the PR-numbered file, and add new
coverage along the way (non-assistant message_end, no-open-span no-ops, status
subcommand, flush-failure path, re-entrant agent_start sweep). test-support.ts is
excluded from the published package via the files allowlist. 141 tests pass.

(Xinwei review comment.)
